+++ monitor-custom/R/sharedUtils/setMonitorIDs.R
@@ -62,4 +62,5 @@ setMonitorIDs <- function(infoList) {
   }
 
   return(infoList)
+  
 }